+++ CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.0.3] - 2021-06-11
+
+### Added
+
+- Add readme files for the [cloudinit](cloudinit) folder
+- Add example for a WLS 12c [host_wls12c](examples/host_wls12c/README.md)
+
 ## [0.0.2] - 2021-06-11
 
 ### Added
@@ -36,3 +43,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [unreleased]: https://github.com/Trivadis/terraform-oci-tvdlab-host
 [0.0.1]: https://github.com/Trivadis/terraform-oci-tvdlab-host/releases/tag/v0.0.1
 [0.0.2]: https://github.com/Trivadis/terraform-oci-tvdlab-host/releases/tag/v0.0.2
+[0.0.3]: https://github.com/Trivadis/terraform-oci-tvdlab-host/releases/tag/v0.0.3